requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
anyio==4.13.0
APScheduler==3.11.2
attrs==26.1.0
certifi==2026.2.25
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
propcache==0.4.1
python-dotenv==1.2.2
python-telegram-bot==22.7
tzdata==2026.1
tzlocal==5.3.1
yarl==1.23.0
flask==3.0.3
